Unit test and some code re-org
