+++ go.mod
@@ -3,7 +3,6 @@ module github.com/dell/csi-metadata-retriever
 go 1.18
 
 require (
-	github.com/container-storage-interface/spec v1.5.0
 	github.com/dell/gocsi v1.5.1
 	github.com/sirupsen/logrus v1.6.0
 	golang.org/x/net v0.0.0-20220127200216-cd36cc0744dd
@@ -17,6 +16,7 @@ require (
 	github.com/PuerkitoBio/purell v1.1.1 // indirect
 	github.com/PuerkitoBio/urlesc v0.0.0-20170810143723-de5bf2ad4578 // indirect
 	github.com/akutz/gosync v0.1.0 // indirect
+	github.com/container-storage-interface/spec v1.5.0 // indirect
 	github.com/coreos/go-semver v0.3.0 // indirect
 	github.com/coreos/go-systemd/v22 v22.3.2 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
@@ -36,6 +36,7 @@ require (
 	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
 	github.com/modern-go/reflect2 v1.0.2 // indirect
 	github.com/munnerz/goautoneg v0.0.0-20191010083416-a7dc8b61c822 // indirect
+	github.com/onsi/ginkgo v1.16.5 // indirect
 	go.etcd.io/etcd/api/v3 v3.5.0 // indirect
 	go.etcd.io/etcd/client/pkg/v3 v3.5.0 // indirect
 	go.etcd.io/etcd/client/v3 v3.5.0 // indirect
